refactor: update pkg name & add caching folder
