+++ packages/utility/file-download/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@autonomys/file-caching",
+  "name": "@autonomys/file-download",
   "packageManager": "yarn@4.7.0",
-  "version": "1.5.13",
+  "version": "1.5.12",
   "license": "MIT",
   "repository": {
     "type": "git",
@@ -45,9 +45,9 @@
     "typescript": "^5.8.3"
   },
   "dependencies": {
-    "@autonomys/asynchronous": "^1.5.13",
-    "@autonomys/auto-dag-data": "^1.5.13",
-    "@autonomys/auto-utils": "^1.5.13",
+    "@autonomys/asynchronous": "^1.5.12",
+    "@autonomys/auto-dag-data": "^1.5.12",
+    "@autonomys/auto-utils": "^1.5.12",
     "@keyvhq/sqlite": "^2.1.7",
     "cache-manager": "^6.4.2",
     "jszip": "^3.10.1",
