+++ packages/utility/file-download/src/caching/index.ts
@@ -1,6 +1,3 @@
-import packageJson from '../package.json'
-
-export const version = packageJson.version
 export * from './defaultConfigs'
 export * from './fileCache'
 export * from './models'
