+++ packages/utility/file-download/src/caching/models.ts
