+++ packages/utility/file-download/src/caching/utils.ts
