+++ packages/utility/file-download/src/index.ts
@@ -1,0 +1,4 @@
+import packageJson from '../package.json'
+
+export const version = packageJson.version
+export * from './caching'
